+++ InnerEye/ML/baselines_util.py
@@ -3,26 +3,37 @@
 #  Licensed under the MIT License (MIT). See LICENSE in the repo root for license information.
 #  ------------------------------------------------------------------------------------------
 import logging
+import os
+import shutil
+import tempfile
 from dataclasses import dataclass
 from pathlib import Path
-from typing import Dict, List, Tuple
+from typing import Dict, List, Optional, Tuple
 
 import matplotlib.pyplot as plt
 import pandas as pd
+from azureml.core import Run
 
 from InnerEye.Azure.azure_config import AzureConfig
-from InnerEye.Azure.azure_util import AZUREML_RUN_FOLDER_PREFIX, get_comparison_baseline_paths, strip_prefix
+from InnerEye.Azure.azure_util import AZUREML_RUN_FOLDER_PREFIX, PARENT_RUN_CONTEXT, RUN_CONTEXT, \
+    get_comparison_baseline_paths, is_offline_run_context, strip_prefix
 from InnerEye.Common import common_util
 from InnerEye.Common.Statistics import wilcoxon_signed_rank_test
 from InnerEye.Common.Statistics.wilcoxon_signed_rank_test import WilcoxonTestConfig
-from InnerEye.Common.common_util import BASELINE_WILCOXON_RESULTS_FILE, ENSEMBLE_SPLIT_NAME, \
-    EPOCH_FOLDER_NAME_PATTERN, FULL_METRICS_DATAFRAME_FILE, ModelProcessing, OTHER_RUNS_SUBDIR_NAME, \
-    SUBJECT_METRICS_FILE_NAME, remove_file_or_directory
+from InnerEye.Common.common_util import BASELINE_WILCOXON_RESULTS_FILE, FULL_METRICS_DATAFRAME_FILE, ModelProcessing, \
+    SUBJECT_METRICS_FILE_NAME, get_best_epoch_results_path, remove_file_or_directory
 from InnerEye.Common.fixed_paths import DEFAULT_AML_UPLOAD_DIR
 from InnerEye.ML.common import DATASET_CSV_FILE_NAME, ModelExecutionMode
 from InnerEye.ML.config import SegmentationModelBase
 from InnerEye.ML.visualizers.metrics_scatterplot import write_to_scatterplot_directory
 from InnerEye.ML.visualizers.plot_cross_validation import convert_rows_for_comparisons, may_write_lines_to_file
+
+REGRESSION_TEST_OUTPUT_FOLDER = "OUTPUT"
+REGRESSION_TEST_AZUREML_FOLDER = "AZUREML_OUTPUT"
+REGRESSION_TEST_AZUREML_PARENT_FOLDER = "AZUREML_PARENT_OUTPUT"
+CONTENTS_MISMATCH = "Contents mismatch"
+MISSING_FILE = "Missing"
+TEXT_FILE_SUFFIXES = [".txt", ".csv", ".json", ".html", ".md"]
 
 
 @dataclass
@@ -53,27 +64,18 @@ def compare_scores_against_baselines(model_config: SegmentationModelBase, azure_
     the Wilcoxon results file.
     """
     # The attribute will only be present for a segmentation model; and it might be None or empty even for that.
-    comparison_blob_storage_paths = getattr(model_config, 'comparison_blob_storage_paths')
+    comparison_blob_storage_paths = model_config.comparison_blob_storage_paths
     if not comparison_blob_storage_paths:
         return
-    outputs_path = model_config.outputs_folder
-    if model_proc == ModelProcessing.ENSEMBLE_CREATION:
-        outputs_path = outputs_path / OTHER_RUNS_SUBDIR_NAME / ENSEMBLE_SPLIT_NAME
-    model_epoch_paths = sorted(outputs_path.glob(EPOCH_FOLDER_NAME_PATTERN))
-    if not model_epoch_paths:
-        logging.warning("Cannot compare scores against baselines: no matches found for "
-                        f"{outputs_path}/{EPOCH_FOLDER_NAME_PATTERN}")
-        return
-    # Use the last (highest-numbered) epoch path for the current run.
-    model_epoch_path = model_epoch_paths[-1]
-    model_metrics_path = model_epoch_path / ModelExecutionMode.TEST.value / SUBJECT_METRICS_FILE_NAME
-    model_dataset_path = model_epoch_path / ModelExecutionMode.TEST.value / DATASET_CSV_FILE_NAME
+    outputs_path = model_config.outputs_folder / get_best_epoch_results_path(ModelExecutionMode.TEST, model_proc)
+    if not outputs_path.is_dir():
+        raise FileNotFoundError(f"Cannot compare scores against baselines: no best epoch results found at {outputs_path}")
+    model_metrics_path = outputs_path / SUBJECT_METRICS_FILE_NAME
+    model_dataset_path = outputs_path / DATASET_CSV_FILE_NAME
     if not model_dataset_path.exists():
-        logging.warning(f"Not comparing with baselines because no {model_dataset_path} file found for this run")
-        return
+        raise FileNotFoundError(f"Not comparing with baselines because no {model_dataset_path} file found for this run")
     if not model_metrics_path.exists():
-        logging.warning(f"Not comparing with baselines because no {model_metrics_path} file found for this run")
-        return
+        raise FileNotFoundError(f"Not comparing with baselines because no {model_metrics_path} file found for this run")
     model_metrics_df = pd.read_csv(model_metrics_path)
     model_dataset_df = pd.read_csv(model_dataset_path)
     comparison_result = download_and_compare_scores(outputs_path,
@@ -173,3 +175,116 @@ def get_comparison_baselines(outputs_folder: Path, azure_config: AzureConfig,
         else:
             raise ValueError(f"could not find comparison data for run {run_rec_id}")
     return comparison_baselines
+
+
+def compare_files(expected: Path, actual: Path) -> str:
+    """
+    Compares two individual files for regression testing. It returns an empty string if the two files appear identical.
+    If the files are not identical, an error message with details is return. This handles known text file formats,
+    where it ignores differences in line breaks. All other files are treated as binary, and compared on a byte-by-byte
+    basis.
+    :param expected: A file that contains the expected contents. The type of comparison (text or binary) is chosen
+    based on the extension of this file.
+    :param actual: A file that contains the actual contents.
+    :return: An empty string if the files appear identical, or otherwise an error message with details.
+    """
+
+    def print_lines(prefix: str, lines: List[str]) -> None:
+        count = 5
+        logging.debug(f"{prefix} {len(lines)} lines, first {count} of those:")
+        logging.debug(os.linesep.join(lines[:count]))
+
+    if expected.suffix in TEXT_FILE_SUFFIXES:
+        # Compare line-by-line to avoid issues with line separators
+        expected_lines = expected.read_text().splitlines()
+        actual_lines = actual.read_text().splitlines()
+        if expected_lines != actual_lines:
+            print_lines("Expected", expected_lines)
+            print_lines("Actual", actual_lines)
+            return CONTENTS_MISMATCH
+    else:
+        expected_binary = expected.read_bytes()
+        actual_binary = actual.read_bytes()
+        if expected_binary != actual_binary:
+            logging.debug(f"Expected {len(expected_binary)} bytes, actual {len(actual_binary)} bytes")
+            return CONTENTS_MISMATCH
+    return ""
+
+
+def compare_folder_contents(expected_folder: Path,
+                            actual_folder: Optional[Path] = None,
+                            run: Optional[Run] = None) -> List[str]:
+    """
+    Compares a set of files in a folder, against files in either the other folder or files stored in the given
+    AzureML run. Each file that is present in the "expected" folder must be also present in the "actual" folder
+    (or the AzureML run), with exactly the same contents, in the same folder structure.
+    For example, if there is a file "<expected>/foo/bar/contents.txt", then there must also be a file
+    "<actual>/foo/bar/contents.txt"
+    :param expected_folder: A folder with files that are expected to be present.
+    :param actual_folder: The output folder with the actually produced files.
+    :param run: An AzureML run
+    :return: A list of human readable error messages, with message and file path. If no errors are found, the list is
+    empty.
+    """
+    messages = []
+    if run and is_offline_run_context(run):
+        logging.warning("Skipping file comparison because the given run context is an AzureML offline run.")
+        return []
+    files_in_run: List[str] = run.get_file_names() if run else []
+    temp_folder = Path(tempfile.mkdtemp()) if run else None
+    for file in expected_folder.rglob("*"):
+        # rglob also returns folders, skip those
+        if file.is_dir():
+            continue
+        # All files stored in AzureML runs use Linux-style path
+        file_relative = file.relative_to(expected_folder).as_posix()
+        if actual_folder:
+            actual_file = actual_folder / file_relative
+        elif temp_folder is not None and run is not None:
+            actual_file = temp_folder / file_relative
+            if file_relative in files_in_run:
+                run.download_file(name=str(file_relative), output_file_path=str(actual_file))
+        else:
+            raise ValueError("One of the two arguments run, actual_folder must be provided.")
+        message = compare_files(expected=file, actual=actual_file) if actual_file.exists() else MISSING_FILE
+        if message:
+            messages.append(f"{message}: {file_relative}")
+        logging.info(f"File {file_relative}: {message or 'OK'}")
+    if temp_folder:
+        shutil.rmtree(temp_folder)
+    return messages
+
+
+def compare_folders_and_run_outputs(expected: Path, actual: Path) -> None:
+    """
+    Compares the actual set of run outputs in the `actual` folder against an expected set of files in the `expected`
+    folder. The `expected` folder can have two special subfolders AZUREML_OUTPUT and AZUREML_PARENT_OUTPUT, that
+    contain files that are expected to be present in the AzureML run context of the present run (AZUREML_OUTPUT)
+    or the run context of the parent run (AZUREML_PARENT_OUTPUT).
+    If a file is missing, or does not have the expected contents, an exception is raised.
+    :param expected: A folder with files that are expected to be present.
+    :param actual: The output folder with the actually produced files.
+    """
+    if not expected.is_dir():
+        raise ValueError(f"Folder with expected files does not exist: {expected}")
+    logging.debug(f"Current working directory: {Path.cwd()}")
+    messages = []
+    for (subfolder, message_prefix, actual_folder, run_to_compare) in \
+            [(REGRESSION_TEST_OUTPUT_FOLDER, "run output files", actual, None),
+             (REGRESSION_TEST_AZUREML_FOLDER, "AzureML outputs in present run", None, RUN_CONTEXT),
+             (REGRESSION_TEST_AZUREML_PARENT_FOLDER, "AzureML outputs in parent run", None, PARENT_RUN_CONTEXT)]:
+        folder = expected / subfolder
+        if folder.is_dir():
+            logging.info(f"Comparing results in {folder} against {message_prefix}:")
+            if actual_folder is None and run_to_compare is None:
+                raise ValueError(f"The set of expected test results in {expected} contains a folder "
+                                 f"{subfolder}, but there is no (parent) run to compare against.")
+            new_messages = compare_folder_contents(folder, actual_folder=actual_folder, run=run_to_compare)
+            if new_messages:
+                messages.append(f"Issues in {message_prefix}:")
+                messages.extend(new_messages)
+        else:
+            logging.info(f"Folder {subfolder} not found, skipping comparison against {message_prefix}.")
+    if messages:
+        raise ValueError(f"Some expected files were missing or did not have the expected contents:{os.linesep}"
+                         f"{os.linesep.join(messages)}")