+++ InnerEye/ML/configs/other/HelloContainer.py
@@ -8,11 +8,12 @@ from typing import Any, Dict, List, Tuple
 import numpy as np
 import torch
 from pytorch_lightning import LightningDataModule, LightningModule
+from pytorch_lightning.metrics import MeanAbsoluteError
 from torch.optim import Adam, Optimizer
 from torch.optim.lr_scheduler import StepLR, _LRScheduler
 from torch.utils.data import DataLoader, Dataset
 
-from InnerEye.Common import fixed_paths_for_tests
+from InnerEye.Common import fixed_paths
 from InnerEye.ML.lightning_container import LightningContainer
 
 
@@ -20,6 +21,7 @@ class HelloDataset(Dataset):
     """
     A simple 1dim regression task, read from a data file stored in the test data folder.
     """
+
     # Creating the data file:
     # import numpy as np
     # import torch
@@ -28,7 +30,7 @@ class HelloDataset(Dataset):
     # x = torch.rand((N, 1)) * 10
     # y = 0.2 * x + 0.1 * torch.randn(x.size())
     # xy = torch.cat((x, y), dim=1)
-    # np.savetxt("Tests/ML/test_data/hellocontainer.csv", xy.numpy(), delimiter=",")
+    # np.savetxt("InnerEye/ML/configs/other/hellocontainer.csv", xy.numpy(), delimiter=",")
     def __init__(self, root_folder: Path, start_index: int, end_index: int) -> None:
         """
         Creates the 1-dim regression dataset.
@@ -37,7 +39,7 @@ class HelloDataset(Dataset):
         :param end_index: The last row to read (exclusive)
         """
         super().__init__()
-        raw_data = np.loadtxt(root_folder / "hellocontainer.csv", delimiter=",")[start_index:end_index]
+        raw_data = np.loadtxt(str(root_folder / "hellocontainer.csv"), delimiter=",")[start_index:end_index]
         self.data = torch.tensor(raw_data, dtype=torch.float)
 
     def __len__(self) -> int:
@@ -51,6 +53,7 @@ class HelloDataModule(LightningDataModule):
     """
     A data module that gives the training, validation and test data for a simple 1-dim regression task.
     """
+
     def __init__(self, root_folder: Path) -> None:
         super().__init__()
         self.train = HelloDataset(root_folder, start_index=0, end_index=50)
@@ -71,10 +74,12 @@ class HelloRegression(LightningModule):
     """
     A simple 1-dim regression model.
     """
+
     def __init__(self) -> None:
         super().__init__()
         self.model = torch.nn.Linear(in_features=1, out_features=1, bias=True)
         self.test_mse: List[torch.Tensor] = []
+        self.test_mae = MeanAbsoluteError()
 
     def forward(self, x: torch.Tensor) -> torch.Tensor:  # type: ignore
         """
@@ -99,7 +104,8 @@ class HelloRegression(LightningModule):
         self.log("loss", loss, on_epoch=True, on_step=False)
         return loss
 
-    def validation_step(self, batch: Dict[str, torch.Tensor], *args: Any, **kwargs: Any) -> torch.Tensor:  # type: ignore
+    def validation_step(self, batch: Dict[str, torch.Tensor], *args: Any,  # type: ignore
+                        **kwargs: Any) -> torch.Tensor:
         """
         This method is part of the standard PyTorch Lightning interface. For an introduction, please see
         https://pytorch-lightning.readthedocs.io/en/stable/starter/converting.html
@@ -142,6 +148,7 @@ class HelloRegression(LightningModule):
         test set (that is done in the test_step).
         """
         self.test_mse = []
+        self.test_mae.reset()
 
     def test_step(self, batch: Dict[str, torch.Tensor], batch_idx: int) -> torch.Tensor:  # type: ignore
         """
@@ -153,8 +160,15 @@ class HelloRegression(LightningModule):
         :param batch_idx: The index (0, 1, ...) of the batch when the data loader is enumerated.
         :return: The loss on the test data.
         """
-        loss = self.shared_step(batch)
+        input = batch["x"]
+        target = batch["y"]
+        prediction = self.forward(input)
+        # This illustrates two ways of computing metrics: Using standard torch
+        loss = torch.nn.functional.mse_loss(prediction, target)
         self.test_mse.append(loss)
+        # Metrics computed using PyTorch Lightning objects. Note that these will, by default, attempt
+        # to synchronize across GPUs.
+        self.test_mae.update(preds=prediction, target=target)
         return loss
 
     def on_test_epoch_end(self) -> None:
@@ -166,6 +180,7 @@ class HelloRegression(LightningModule):
         """
         average_mse = torch.mean(torch.stack(self.test_mse))
         Path("test_mse.txt").write_text(str(average_mse.item()))
+        Path("test_mae.txt").write_text(str(self.test_mae.compute()))
 
 
 class HelloContainer(LightningContainer):
@@ -176,9 +191,10 @@ class HelloContainer(LightningContainer):
     You can train this model by running `python InnerEye/ML/runner.py --model=HelloContainer` on the local box,
     or via `python InnerEye/ML/runner.py --model=HelloContainer --azureml=True` in AzureML
     """
+
     def __init__(self) -> None:
         super().__init__()
-        self.local_dataset = fixed_paths_for_tests.full_ml_test_data_path()
+        self.local_dataset = fixed_paths.repository_root_directory() / "InnerEye" / "ML" / "configs" / "other"
         self.num_epochs = 20
 
     # This method must be overridden by any subclass of LightningContainer. It returns the model that you wish to
@@ -196,7 +212,8 @@ class HelloContainer(LightningContainer):
     # training, and cook them into a nice looking report. Here, the report is a simple text file.
     def create_report(self) -> None:
         # This just prints out the test MSE, but you could also generate a Jupyter notebook here, for example.
-        test_mse = float(Path("test_mse.txt").read_text())
-        report = f"Performance on test set: MSE = {test_mse}"
+        test_mse = Path("test_mse.txt").read_text().strip()
+        test_mae = Path("test_mae.txt").read_text().strip()
+        report = f"Performance on test set: MSE = {test_mse}, MAE = {test_mae}"
         print(report)
         Path("report.txt").write_text(report)